backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Nicolas Caicedo",email = "nicolas.caicedo1799@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "alembic (>=1.15.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
pytest-cov = "^6.0.0"
faker = "^37.0.2"

